architecture-examples/backend/requirements.txt
fastapi>=0.104.0
strawberry-graphql[fastapi]>=0.215.0
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0
httpx>=0.25.0
pydantic-settings>=2.0.0

# Data & Analytics
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
plotly>=5.15.0

# Database & Caching
redis>=5.0.0
sqlalchemy>=2.0.0
alembic>=1.12.0
psycopg2-binary>=2.9.7

# Background Tasks
celery>=5.3.0

# Crypto APIs & WebSocket
websockets>=11.0.0
aioredis>=2.0.0

# Development & Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Security & Auth
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
PyJWT>=2.8.0
bcrypt>=4.0.0